requirements.txt
autopep8==1.5.5
certifi==2020.12.5
chardet==3.0.4
idna==2.5
numpy==1.20.1
pandas==1.2.3
pycodestyle==2.6.0
python-dateutil==2.8.1
pytz==2021.1
requests2==2.16.0
six==1.15.0
toml==0.10.2
urllib3==1.21.1
